Abstract 'possibly bare' repository check to its own file
