+++ packages/expo-doctor/src/checks/ProjectSetupCheck.ts
@@ -1,9 +1,9 @@
-import spawnAsync from '@expo/spawn-async';
 import glob from 'fast-glob';
 import fs from 'fs';
 import path from 'path';
 
 import { DoctorCheck, DoctorCheckParams, DoctorCheckResult } from './checks.types';
+import { isFileIgnoredAsync } from '../utils/isFileIgnoredAsync';
 
 export class ProjectSetupCheck implements DoctorCheck {
   description = 'Check for common project setup issues';
@@ -34,19 +34,6 @@ export class ProjectSetupCheck implements DoctorCheck {
       }
     }
 
-    // ** possibly-unintentionally-bare check **
-
-    if (
-      exp.plugins?.length &&
-      // git check-ignore needs a specific file to check gitignore, we choose Podfile
-      ((await existsAndIsNotIgnoredAsync(path.join(projectRoot, 'ios', 'Podfile'))) ||
-        (await existsAndIsNotIgnoredAsync(path.join(projectRoot, 'android', 'Podfile'))))
-    ) {
-      issues.push(
-        'This project has native project folders but also has config plugins, indicating it is configured to use Prebuild. EAS Build will not sync your native configuration if the ios or android folders are present. Add these folders to your .gitignore file if you intend to use prebuild (aka "managed" workflow).'
-      );
-    }
-
     // ** multiple lock file check **
 
     const lockfileCheckResults = await Promise.all(
@@ -72,25 +59,6 @@ export class ProjectSetupCheck implements DoctorCheck {
       issues,
     };
   }
-}
-
-async function existsAndIsNotIgnoredAsync(filePath: string): Promise<boolean> {
-  return fs.existsSync(filePath) && !(await isFileIgnoredAsync(filePath));
-}
-
-async function isFileIgnoredAsync(filePath: string): Promise<boolean> {
-  try {
-    await spawnAsync('git', ['check-ignore', '-q', filePath], {
-      cwd: path.normalize(await getRootPathAsync()),
-    });
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-async function getRootPathAsync(): Promise<string> {
-  return (await spawnAsync('git', ['rev-parse', '--show-toplevel'])).stdout.trim();
 }
 
 async function areAnyMatchingPathsIgnoredAsync(filePath: string): Promise<boolean> {
